现金支付完成

+++ mmp-services/mmp-services-charge/src/main/java/com/aaa/charge/vo/PostObjVo2.java
@@ -2,10 +2,12 @@ package com.aaa.charge.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PostObjVo2 {
     //总金额
-    private Double orderAmount;
+    private BigDecimal orderAmount;
     //病历ID
     private String chId;
     //患者名字
